requirements.txt
Flask
ultralytics
torch
torchvision
Pillow